Figuring out how to do this.
