+++ src/tiles/Bedrock.ts
@@ -1,0 +1,7 @@
+export default {
+  BEDROCK: {
+    name: 'BEDROCK',
+    texture: '/textures/pixel/bedrock-2.png',
+    symbol: '#',
+  },
+}
