+++ src/tiles/Boulder.ts
@@ -1,0 +1,18 @@
+export default {
+  DIRT_BOULDER: {
+    name: 'DIRT_BOULDER',
+    texture: '/textures/pixel/dirt-boulder.png',
+    symbol: 'O',
+  },
+
+  BEDROCK_BOULDER: {
+    name: 'BEDROCK_BOULDER',
+    texture: '/textures/pixel/bedrock-boulder.png',
+    symbol: 'o',
+  },
+
+  FALLING_BOULDER: {
+    name: 'FALLING_BOULDER',
+    texture: '/textures/pixel/bedrock-boulder.png',
+  },
+}
