+++ src/tiles/Diamond.ts
@@ -1,0 +1,13 @@
+export default {
+  DIRT_DIAMOND: {
+    name: 'DIRT_DIAMOND',
+    texture: '/textures/pixel/dirt-diamond.png',
+    symbol: 'D',
+  },
+
+  BEDROCK_DIAMOND: {
+    name: 'BEDROCK_DIAMOND',
+    texture: '/textures/pixel/bedrock-diamond.png',
+    symbol: 'd',
+  },
+}
