+++ src/tiles/Dirt.ts
@@ -1,0 +1,7 @@
+export default {
+  DIRT: {
+    name: 'DIRT',
+    texture: '/textures/pixel/dirt.png',
+    symbol: '.',
+  },
+}
