+++ src/tiles/Explosion.ts
@@ -1,0 +1,6 @@
+export default {
+  EXPLOSION: {
+    name: 'EXPLOSION',
+    texture: '/textures/pixel/boom.gif',
+  },
+}
