+++ src/tiles/Finish.ts
@@ -1,0 +1,7 @@
+export default {
+  FINISH: {
+    name: 'FINISH',
+    texture: '/textures/pixel/finish.gif',
+    symbol: 'f',
+  },
+}
