+++ src/tiles/Nothing.ts
@@ -1,0 +1,7 @@
+export default {
+  NOTHING: {
+    name: 'NOTHING',
+    texture: '/textures/pixel/bedrock.png',
+    symbol: ' ',
+  },
+}
