+++ src/tiles/Player.ts
@@ -1,0 +1,7 @@
+export default {
+  PLAYER: {
+    name: 'PLAYER',
+    texture: '/textures/pixel/player.gif',
+    symbol: 'p',
+  },
+}
